Updated structures of EvoData so that we split it up into sensible sub-structures.
Collected the shared structures of SL-PDMP and CABPS (Lagrangian/Ver 6.2) into single files to avoid code replication. This was done in anticipation of fully-split methods that will share a subset of EvoData with SL-PDMP and CABPS.

Tested all implemented methods on TwinPeaks, runs fine. Further testing needed to verify integrity and type-stability.